Add shell script files
Add shell script files for downloading the NIH dbSNP data for hg19 and hg38 genomes including preparing it for clean up and conversion to MCF format. Also, adds shell script that downloads the NCBI Chromosome Assembly files for hg19 and hg38 and creates a text file that is a key for converting chromosome RefSeq IDs to the corresponding dcid.